Dockerized the almighty-ui
I've added a Dockefile that will be used for building

the almightyui/almighty-ui image on docker hub:

https://hub.docker.com/r/almightyui/almighty-ui/

The package.json has a new script that is explicitly used
in the Dockerfile's command (CMD). This was done so that
we can handle what gets executed once the container is
started separately from the rest of the scripts.